[DK-48] refactor:  인터페이스 삭제
